refactor: correct board style
